fix warning in Debug mode

+++ src/fsparser.cpp
@@ -185,7 +185,7 @@ int FsParser::parse()
 	/* For configuration data checksum: number of lines can't be higher than
 	 * the number indicates in TN653 but may be smaller (seen with the GW1NS-2C).
 	 */
-	if (stoi(_hdr["ConfDataLength"]) < nb_line)
+	if (stoul(_hdr["ConfDataLength"]) < nb_line)
 		nb_line = stoi(_hdr["ConfDataLength"]);
 
 	/* drop now useless header */
